chore(client): rename `App` => `Home`
